requirements.txt
dateparser==1.1.2
discord.py==2.0.1
inflect==6.0.2
lol_id_tools==2.0.1
matplotlib==3.6.1
mplcyberpunk==0.6.0
pytest==7.1.3
rapidfuzz==2.11.1
SQLAlchemy==1.4.42
tabulate==0.9.0
types-tabulate==0.9.0.0
trueskill==0.4.5
psycopg2==2.9.4
git+https://github.com/Rapptz/discord-ext-menus
alembic==1.8.1
pyot==6.0.5
fastapi==0.85.1
uvicorn==0.19.0
black==22.10.0
mypy==0.982
# remove with sqlalchemy 2.0
sqlalchemy2-stubs==0.0.2a29
